requirements.txt
discord.py
youtube_dl
beautifulsoup4
imageio
pillow
pyyaml
pymongo
peewee
pynacl
